+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/content/FileNamer.kt
@@ -20,6 +20,7 @@ package org.wycliffeassociates.otter.common.domain.content
 
 import org.wycliffeassociates.otter.common.audio.AudioFileFormat
 import org.wycliffeassociates.otter.common.data.primitives.ContentType
+import java.util.regex.Pattern
 
 class FileNamer(
     val start: Int? = null,
@@ -95,5 +96,14 @@ class FileNamer(
 
     companion object {
         const val DEFAULT_RC_SLUG = "reg"
+        val takeFilenamePattern: Pattern = run {
+            val chapter = """_c(\d+)"""
+            val verse = """(?:_v(\d+))?"""
+            val sort = """(?:_s(\d+))?"""
+            val type = """(?:_([A-Za-z]+))?"""
+            val take = """_t(\d+)"""
+            val extensionDelim = """\."""
+            Pattern.compile(chapter + verse + sort + type + take + extensionDelim)
+        }
     }
 }
