+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/project/exporter/IDirectoryExporter.kt
@@ -1,4 +1,3 @@
 package org.wycliffeassociates.otter.common.domain.project.exporter
 
-interface IDirectoryExporter : ProjectExporter {
-}
+interface IDirectoryExporter : IProjectExporter
