--- common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/project/exporter/ProjectExporterCallback.kt
@@ -1,4 +1,0 @@
-package org.wycliffeassociates.otter.common.domain.project.exporter
-
-interface ProjectExporterCallback {
-}
